adding top-level README.md and links to slides
Now that we actually have this repo generating slides via tags and GHA
we can refer to the slides in the top-level README.md. Additionally, Ive
included some info about what the repo is for and what tools I use
during a presentation.

Included the present devShell for pdfpc which I need to get better about
leveraging all of its slick features for my next talk.